asynchronous/embassy/task: Back-off interrupt checking in certain cases
The interrupt line may continue to be asserted after interrupts have
been checked. This is expected if we currently have interrupts disabled
or may happen if interrupt processing encountered an error. When this
happens the task can starve other tasks because `wait_for_low` will
immediately return. Introduce a to prevent this happening.